Add readme

+++ build.ps1
@@ -5,9 +5,11 @@ $loginAppDir = Join-Path $baseLocation "src\HCGStudio.TheLauncherLogin"
 $appDir = Join-Path $baseLocation "\src\HCGStudio.TheLauncher"
 $outputDir = Join-Path $baseLocation "build"
 
-Remove-Item $outputDir -Recurse
-mkdir $outputDir
+if(Test-Path $outputDir){
+    Remove-Item $outputDir -Recurse
+}
 
+mkdir $outputDir
 
 Set-Location $loginAppDir
 msbuild -t:restore
@@ -24,6 +26,6 @@ Copy-Item (Join-Path $loginAppBuildResult "*") $outputDir -Recurse
 #Build Native
 Set-Location $appDir
 dotnet publish -r win-x64 -c release
-Copy-Item (Join-Path $appDir "bin\Release\net5.0\win-x64\publish" "*.exe") $outputDir -Recurse
+Copy-Item (Join-Path $appDir "bin\x64\release\net5.0\win-x64\publish\*.exe") $outputDir -Recurse
 
 Set-Location $baseLocation
